Assignment 2 Updated
