+++ DungeonExplorer.sln
@@ -1,9 +1,8 @@
-﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.11.35222.181
+VisualStudioVersion = 17.5.2.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DungeonExplorer", "DungeonExplorer.csproj", "{FE35A2D4-F16B-43F2-A317-D033ABD5B6B3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DungeonExplorer", "DungeonExplorer.csproj", "{19D6923E-D34D-1A9D-F5E0-4D7EE1E57E1E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +10,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FE35A2D4-F16B-43F2-A317-D033ABD5B6B3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FE35A2D4-F16B-43F2-A317-D033ABD5B6B3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FE35A2D4-F16B-43F2-A317-D033ABD5B6B3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FE35A2D4-F16B-43F2-A317-D033ABD5B6B3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{19D6923E-D34D-1A9D-F5E0-4D7EE1E57E1E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{19D6923E-D34D-1A9D-F5E0-4D7EE1E57E1E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{19D6923E-D34D-1A9D-F5E0-4D7EE1E57E1E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{19D6923E-D34D-1A9D-F5E0-4D7EE1E57E1E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {82C9D992-5C01-4E3B-B552-64DE84B30EAF}
+		SolutionGuid = {85987B5E-E89F-426F-AC79-2E8ECFECDB87}
 	EndGlobalSection
 EndGlobal
