--- DungeonExplorer/DungeonExplorer.sln
@@ -1,24 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.2.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DungeonExplorer", "DungeonExplorer.csproj", "{19D6923E-D34D-1A9D-F5E0-4D7EE1E57E1E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{19D6923E-D34D-1A9D-F5E0-4D7EE1E57E1E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{19D6923E-D34D-1A9D-F5E0-4D7EE1E57E1E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{19D6923E-D34D-1A9D-F5E0-4D7EE1E57E1E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{19D6923E-D34D-1A9D-F5E0-4D7EE1E57E1E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {85987B5E-E89F-426F-AC79-2E8ECFECDB87}
-	EndGlobalSection
-EndGlobal
