+++ GoblinDragon.cs
