+++ Program.cs
@@ -1,20 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using System;
+using DungeonExplorer_Simplified;  // Add the necessary 'using' directive
 
-namespace DungeonExplorer
+namespace DungeonExplorer_Simplified
 {
-    internal class Program
+    class Program
     {
         static void Main(string[] args)
         {
-            Game game = new Game();
+            Game game = new Game(); // Ensure Game class is available
             game.Start();
-            Console.WriteLine("Waiting for your Implementation");
-            Console.WriteLine("Press any key to exit...");
-            Console.ReadKey();
         }
     }
 }
